ingestor/requirements.txt
requests
beautifulsoup4
boto3
python-dotenv
lxml
